% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/american_to_odds.R
\name{american_to_odds}
\alias{american_to_odds}
\title{Convert American betting line to odds}
\usage{
american_to_odds(line)
}
\arguments{
\item{line}{An American style betting line}
}
\value{
The proportion of the original wager paid out
}
\description{
Convert American betting line to odds
}
\examples{
american_to_odds(-110)
}
