% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/american_to_odds.R
\name{american_to_prob}
\alias{american_to_prob}
\title{Convert American betting line to probability}
\usage{
american_to_prob(line = 100)
}
\arguments{
\item{line}{integer indicating the team's betting line}
}
\value{
Numeric vector of probabilities derived from home and away lines
}
\description{
Convert American betting line to probability
}
\examples{
american_to_prob(-110, 100, home = TRUE)
}
