% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int_ols.R
\name{conf_int_ols}
\alias{conf_int_ols}
\title{Obtain a confidence interval using OLS}
\usage{
conf_int_ols(X, y, cont, d = 0, a = 0.05)
}
\arguments{
\item{X}{A design matrix}

\item{y}{Values of the outputs}

\item{cont}{Contrasts}

\item{d}{Value of cont*beta under null hypothesis}

\item{a}{Confidence level between 0 and 1}
}
\value{
A data frame with the estimate of cont*beta-d with lower and upper bounds of
the confidence interval
}
\description{
Obtain a confidence interval using OLS
}
\examples{
G <- regssn2021
List <- XY_differences(G)
X <- ListX
y <- List$Y_diff
d <- length(List$teams)
cont <- c(0, rep(0,d))
conf_int_ols(X, y, cont, d = 0, a = 0.1)
}
