% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepData.R
\name{drop_incomplete_games}
\alias{drop_incomplete_games}
\title{Remove games from data set that have missing information}
\usage{
drop_incomplete_games(data, game_ids = FALSE)
}
\arguments{
\item{data}{A data frame with columns for Team, Score, Year, and Week
produced by scrape_games.  These 3 columns must be numeric.}

\item{game_ids}{Logical argument to indicate if we want the returned data frame
to include game_id numbers}
}
\value{
A cleaned version of game data with games that had incomplete data
removed
}
\description{
Remove games from data set that have missing information
}
\examples{
drop_incomplete_games(regssn2021, game_ids = TRUE)
}
