% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eML_ols.R
\name{eML_emp}
\alias{eML_emp}
\title{Calculate the expected value of wagering on either team's moneyline using the
empircal distribution.}
\usage{
eML_emp(
  design,
  home,
  away,
  wager = 1,
  hBL = -110,
  aBL = -110,
  home_effect = TRUE
)
}
\arguments{
\item{design}{List of season data as produced by get_design(), contains
design matrix X, scores Y_diff, and a list of teams}

\item{home}{Name of the home team, could be a substring}

\item{away}{Name of the away team, could be a substring}

\item{wager}{Amount of money being wagered in dollars}

\item{hBL}{The home team's moneyline given as American}

\item{aBL}{The away team's moneyline given as American}

\item{home_effect}{Logical, indicates if there is home field advantage}
}
\value{
A data frame indicating the the expected values from wagering on each
team's moneyline based on the empircal distribution.
}
\description{
Calculate the expected value of wagering on either team's moneyline using the
empircal distribution.
}
\examples{
design <- get_design(regssn2021)
eML_emp(design, "Rams", "Bengals", wager = 100, hBL = -155, aBL = 185)
}
