% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eML.R
\name{eML_ols}
\alias{eML_ols}
\title{Calculate the expected value for wagering on each team assuming a normal
distribution.}
\usage{
eML_ols(
  data,
  home,
  away,
  wager = 1,
  hBL = -110,
  aBL = -110,
  home_effect = TRUE
)
}
\arguments{
\item{data}{List of season data as produced by XY_differences, contains
design matrix X, scores Y_diff, and a list of teams}

\item{home}{Name of the home team, could be a substring}

\item{away}{Name of the away team, could be a substring}

\item{wager}{Amount of money being wagered in dollars}

\item{hBL}{The home team's moneyline given as American}

\item{aBL}{The away team's moneyline given as American}

\item{home_effect}{Logical, indicates if there is home field advantage}
}
\value{
A data frame containing the names of the two teams and the expected
values for wagering on either moneyline.
}
\description{
Calculate the expected value for wagering on each team assuming a normal
distribution.
}
\examples{
G <- regssn2021
data <- XY_differences(G)
eML_ols(data, "Bills", "Patriots", wager = 20, hBL = -160, aBL = 180)
}
