% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSpread_ols.R
\name{eSpread_ols}
\alias{eSpread_ols}
\title{Predict the probability of each team beating the spread using ols estimates}
\usage{
eSpread_ols(
  data,
  home,
  away,
  hspread,
  aspread = NULL,
  home_effect = TRUE,
  wager = 1,
  hBL = -110,
  aBL = -110
)
}
\arguments{
\item{data}{List of season data as produced by XY_differences, contains
design matrix X, scores Y_diff, and a list of teams}

\item{home}{Name of the home team, could be a substring}

\item{away}{Name of the away team, could be a substring}

\item{hspread}{Point spread for the home team}

\item{aspread}{Point spread for the away team}

\item{home_effect}{Logical, indicates if we consider home field advantage}

\item{wager}{Amount of money being wagered in dollars}

\item{hBL}{The betting line for the home team against the spread, American}

\item{aBL}{The betting line for the away team against the spread, American}
}
\value{
A dataframe indicating the expected value of wagering on each team to
beat the spread.
}
\description{
Predict the probability of each team beating the spread using ols estimates
}
\examples{
G <- regssn2021
List <- XY_differences(G)
eSpread_ols(List, "Patriots", "Bills", hspread = -5, hBL = -115, aBL = -105)
}
