% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepData.R
\name{games_wide_to_design}
\alias{games_wide_to_design}
\title{Obtain a design from a wide-fromat dataframe of games}
\usage{
games_wide_to_design(data)
}
\arguments{
\item{data}{Data in a a wide format with columns for home team, home score,
away team, away score, season, week, game type, location, result, and total}
}
\value{
A list with 7 components:  2 design matrices (X, X_sum), and 3 vectors
of observations of season data (Y_diff, Y_sum, Y_binary), the list of teams,
and original dataframe of scores.
}
\description{
Obtain a design from a wide-fromat dataframe of games
}
\details{
This function is intended to help when games are scraped using
the R package \code{nflreadr}.  The data is then pivoted into a wide format and
also includes a column "location" which incidates when the game is played at
a Home or Neutral site.
}
