% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design.R
\name{get_design}
\alias{get_design}
\title{Get design matrix and values from a season}
\usage{
get_design(data)
}
\arguments{
\item{data}{A data frame with columns for Team, Score, Year, and Week
produced by scrape_games.  Last 3 columns must be numeric.}
}
\value{
A list of 9 components, including 2 design matrices (X, X_sum),
and 3 vectors of observations of season data (Y_diff, Y_sum, Y_binary), the
list of teams, original dataframe of scores, and date range of games.
}
\description{
Get design matrix and values from a season
}
\examples{
games <- regssn2021
L <- get_design(games)
}
