% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/american_to_odds.R
\name{logistic}
\alias{logistic}
\title{Compute the logistic}
\usage{
logistic(x)
}
\arguments{
\item{x}{A real number}
}
\value{
A probability
}
\description{
The logistic function is given by
\deqn{\text{logistic}(x) = \frac{1}{1 + \exp(-x)}}
}
\examples{
logistic(0)
}
