% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/american_to_odds.R
\name{logit}
\alias{logit}
\title{Compute the logit}
\usage{
logit(p)
}
\arguments{
\item{p}{A probability between 0 and 1}
}
\value{
The logit of the probability
}
\description{
This is the inverse of the logistic function, given by
\deqn{\text{logit}(p) = \text{log}(p/(1-p))}
}
\examples{
# logit(0.5) is 0
logit(0.5)
}
