% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overunder.R
\name{over_under_ols}
\alias{over_under_ols}
\title{Use ols to estimate the total points scored in a game between tm1 and tm2}
\usage{
over_under_ols(design, tm1, tm2, a = 0.05, verbose = TRUE)
}
\arguments{
\item{design}{Design object with information from a collection of games.
This is a list that is the output of get_design()}

\item{tm1}{A substring of the name of the first team}

\item{tm2}{A substring of the name of the second team}

\item{a}{The confidence level of the interval we estimate}

\item{verbose}{Whether or not a lengthy readout of the estimate is desired}
}
\value{
A dataframe with the estimate and confidence interval
}
\description{
Use ols to estimate the total points scored in a game between tm1 and tm2
}
\examples{
design <- get_design(regssn2021)
over_under_ols(design, "Rams", "Bengals")
}
