% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepData.R
\name{pivot_games_wide}
\alias{pivot_games_wide}
\title{Pivot games data into a wider format}
\usage{
pivot_games_wide(games)
}
\arguments{
\item{games}{A data frame with columns for Team, Score, Year, and Week where
each game is represented by two rows, one for each team}
}
\value{
A data frame with columns for game_id, home_team, home_score,
away_team, and away_score for each game
}
\description{
Pivot games data into a wider format
}
\examples{
games <- regssn2021
pivot_games_wide(games)
}
