% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_spread_ols.R
\name{point_spread_ols}
\alias{point_spread_ols}
\title{Estimate the point spread from a season design object}
\usage{
point_spread_ols(data, home, away, home_effect = TRUE, a = 0.05)
}
\arguments{
\item{data}{A list containing a design matrix, score differences, and teams,
as produced by the XY_differences function}

\item{home}{The name of the home team, may be a substring}

\item{away}{The name of the away team, may be a substring}

\item{home_effect}{Logical, TRUE if we want to estimate with home field advantage}

\item{a}{The level of the confidence interval}
}
\value{
A data frame describing the estimated point spread and the limits of a
confidence level
}
\description{
Estimate the point spread from a season design object
}
