% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regssn2021}
\alias{regssn2021}
\title{Results of games from the 2021 regular season}
\format{
A data frame with 544 rows and 4 variables:
\describe{
\item{Team}{Team playing in the game.  Alternates away, home, away,...}
\item{Score}{Number of points scored}
\item{Year}{The year the game was played}
\item{Week}{The week in the season when the game was played}
}
}
\source{
\url{https://www.pro-football-reference.com/}
}
\usage{
regssn2021
}
\description{
A dataset containing the results of games from the 2021 NFL season
}
\keyword{datasets}
