% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_games.R
\name{scrape_games}
\alias{scrape_games}
\title{scrape_games}
\usage{
scrape_games(ssn = 2021, wk_start = 1, wk_stop)
}
\arguments{
\item{ssn}{The starting year of the season from where we are scraping games}

\item{wk_start}{The first week number of games being scraped}

\item{wk_stop}{The final week number of games being scraped}
}
\value{
A data frame with columns Team, Score, Year, Week
}
\description{
scrape_games
}
\examples{
G <- scrape_games(ssn = 2017, wk_stop = 17)
tail(G)
}
