% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadprob.R
\name{spreadprob_emp}
\alias{spreadprob_emp}
\title{Compute win probability using the empirical distribution}
\usage{
spreadprob_emp(
  design,
  home,
  away,
  hspread = 0,
  aspread = NULL,
  home_effect = TRUE
)
}
\arguments{
\item{design}{List of season data as produced by get_design(), contains
design matrix X, scores Y_diff, and a list of teams}

\item{home}{Name of the home team, could be a substring}

\item{away}{Name of the away team, could be a substring}

\item{hspread}{The listed point spread for the home team}

\item{aspread}{The listed point spread for the away team}

\item{home_effect}{Logical, indicates if there is home field advantage}
}
\value{
A data frame containing the teams and their probabilities of beating
the spread
}
\description{
Compute win probability using the empirical distribution
}
\examples{
design <- get_design(regssn2021)
spreadprob_emp(design, "Patriots", "Bills", hspread = 5)
}
