% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadprob_ols.R
\name{spreadprob_ols}
\alias{spreadprob_ols}
\title{Predict the probability of each team beating the spread using ols estimates}
\usage{
spreadprob_ols(data, home, away, hspread, aspread = NULL, home_effect = TRUE)
}
\arguments{
\item{data}{List of season data as produced by XY_differences, contains
design matrix X, scores Y_diff, and a list of teams}

\item{home}{Name of the home team, could be a substring}

\item{away}{Name of the away team, could be a substring}

\item{hspread}{Point spread for the home team}

\item{aspread}{Point spread for the away team}

\item{home_effect}{Logical, indicates if we consider home field advantage}
}
\value{
A dataframe indicating the probabilities of each team beating the spread
}
\description{
Predict the probability of each team beating the spread using ols estimates
}
