% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{winprob_logistic}
\alias{winprob_logistic}
\title{Use a logistic regression to predict win probability}
\usage{
winprob_logistic(
  data,
  home,
  away,
  home_advantage_fit = TRUE,
  home_advantage_predict = TRUE,
  home_win_prob = TRUE
)
}
\arguments{
\item{data}{List of season data as produced by get_design(), contains}

\item{home}{Character vector of home team names}

\item{away}{Character vector of away team names}

\item{home_advantage_fit}{Logical, TRUE if we want to fit the logistic model
with home field advantage}

\item{home_advantage_predict}{Logical, TRUE if we want our prediction to
include the home field advantage estimated in the logistic model}

\item{home_win_prob}{Logical, TRUE if we want to calculate the home team's win probability,
otherwise we estimate the away team's win probability}
}
\value{
A list containing the predictions, estimates, and the model object
}
\description{
Use a logistic regression to predict win probability
}
\details{
Our logistic model only uses the binary outcome of the home team winning or losing.
Each team effect is estimated as a coefficient in the logistic model, and we
also can estiamte an intercept for the home field advantage.  If we let
\eqn{\beta_i} be the coefficient for team \eqn{i} and \eqn{\gamma} be the
intercept for home field advantage, then our model says the probability of
team \eqn{i} defeating team \eqn{j} is at home is given by
\deqn{ y_{ij} = \frac{\text{exp}(\beta_i - \beta_j + \gamma)}{1 - \text{exp}(\beta_i - \beta_j + \gamma)} }
}
\examples{
#None
}
