% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XY_differences.R
\name{XY_differences}
\alias{XY_differences}
\title{From a collection of game scores, create a design matrix X and a vector of differences Y_diff
to use for linear regression.}
\usage{
XY_differences(games)
}
\arguments{
\item{games}{}
}
\value{
a list
}
\description{
From a collection of game scores, create a design matrix X and a vector of differences Y_diff
to use for linear regression.
}
\examples{
result <- XY_differences(sample)
}
