% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_spread_ols.R
\name{point_spread_ols}
\alias{point_spread_ols}
\title{Estimate the point spread from a season design object}
\usage{
point_spread_ols(
  data,
  home,
  away,
  home_advantage_fit = TRUE,
  home_advantage_predict = TRUE,
  a = 0.05
)
}
\arguments{
\item{data}{A list containing a design matrix, score differences, and teams,
as produced by the XY_differences function}

\item{home}{A vector of home team names, may use distinct substrings}

\item{away}{A vector of away team names, may use distinct substrings}

\item{home_advantage_fit}{Logical, TRUE if we want to fit the OLS model with home field advantage}

\item{home_advantage_predict}{Logical, TRUE if we want to estimate point spreads with home field advantage}

\item{a}{The level of the confidence interval}
}
\value{
A data frame describing the estimated point spread and the limits of a
confidence level
}
\description{
Estimate the point spread from a season design object
}
\details{
The order of teams in home and away must correspond with the games
to predict on, meaning the first element of \code{home} plays the first element of
\code{away}, and so on. The function uses the OLS method to estimate the point
spread and the confidence interval assumes a normal distribution of
point spread.
}
\examples{
G <- regssn2021
List <- XY_differences(G)
point_spread_ols(List, "Patriots", "Bills")
}
