% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadprob.R
\name{spreadprob_normal}
\alias{spreadprob_normal}
\title{Calculate the probability of each team beating the spread assuming normal
errors}
\usage{
spreadprob_normal(
  design,
  home,
  away,
  h_spread,
  a_spread,
  home_effect = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{design}{List of season data as produced by get_design(), contains
design matrix X, scores Y_diff, and a list of teams}

\item{home}{Name of the home team, could be a substring}

\item{away}{Name of the away team, could be a substring}

\item{home_effect}{Logical, indicates if there is home field advantage}

\item{verbose}{Logical, indicates if we want output to read win probabilities}
}
\value{
A data frame containing the teams and their win probabilities
}
\description{
Calculate the probability of each team beating the spread assuming normal
errors
}
\examples{
G <- regssn2021
List <- get_design(G)
spreadprob_normal(List, "Patriots", "Bills")
}
