% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winprob.R
\name{winprob_emp}
\alias{winprob_emp}
\title{Compute win probability using the empirical distribution}
\usage{
winprob_emp(design, home, away, home_effect = TRUE, verbose = TRUE)
}
\arguments{
\item{design}{List of season data as produced by get_design(), contains
design matrix X, scores Y_diff, and a list of teams}

\item{home}{Name of the home team, could be a substring}

\item{away}{Name of the away team, could be a substring}

\item{home_effect}{Logical, indicates if there is home field advantage}

\item{verbose}{Logical, indicates if we want output to read win probabilities}
}
\value{
A data frame containing the teams and their win probabilities
}
\description{
Compute win probability using the empirical distribution
}
\examples{
G <- regssn2021
design <- get_design(G)
winprob_emp(design, "Patriots", "Bills")
}
