% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XY_differences.R
\name{XY_differences}
\alias{XY_differences}
\title{From a collection of game scores, create a design matrix X and a vector of differences Y_diff
to use for linear regression.}
\usage{
XY_differences(data)
}
\arguments{
\item{data}{A data frame with columns for Team, Score, Year, and Week produced by scrape_games.  Last 3 columns must be numeric.}
}
\value{
A list with 6 elements
}
\description{
From a collection of game scores, create a design matrix X and a vector of differences Y_diff
to use for linear regression.
}
\examples{
G <- regssn2021
List <- XY_differences(G)
List$teams
}
